Add HasStatus instance for NoContent
This allows us to easily define APIs that may return a 204 No Content
status. If we want to define an API which returns no content on another
status code, for example 205 Reset Content, we can do WithStatus 205
NoContent which would work as intended.

I think this gets us all the functionality to say we have
UVerbNoContent, although we might like to have a shorthand for something
that responds with no content on many status codes.